refactor(client): move ParallelGet into client package
Move ParallelGet and its helpers from internal/cache into client so it
can drive either the Client impl or the cache.Cache interface via a new
narrow RangeReader interface. The ETagKey constant moves down to client
and is re-exported from cache. A thin cache.ParallelGet wrapper is kept
for existing callers.
